backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
google-cloud-aiplatform>=1.49.0
